+++ ss13_tools/log_downloader/ckey.py
@@ -1,4 +1,4 @@
-from .abstract_downloader import LogDownloader
+from .abstract import LogDownloader
 from .constants import DEFAULT_NUMBER_OF_ROUNDS, DEFAULT_ONLY_PLAYED, DEFAULT_OUTPUT_PATH
 from byond import canonicalize
